// Angular/Hrms/src/model/employee.ts
export class Employee {
    id: number = 0;
    employeeName: string = '';
    employeeDepartment: string = '';
    employeeDesignation: string = '';
    employeeSalary: number = 0;
    assetName: any;
    assetModelNo: any;
    assetType: any;

    //assets: Array<{ AssetId: number; AssetName: string, AssetModelNo: String, AssetType: String }>;
}